Refactored project structure for better organization
