+++ frontend/src/components/ResourceList.jsx
