+++ frontend/src/pages/DashBoard.jsx
@@ -5,11 +5,11 @@ import {
   BrowserRouter,
   Switch,
 } from "react-router-dom";
-import { Outlet } from "react-router-dom";
+
 import "../styles.css";
-import AppBar from "@mui/material/AppBar";
+
 import Button from "@mui/material/Button";
-import CameraIcon from "@mui/icons-material/PhotoCamera";
+
 import Card from "@mui/material/Card";
 import CardActions from "@mui/material/CardActions";
 import CardContent from "@mui/material/CardContent";
@@ -18,12 +18,14 @@ import CssBaseline from "@mui/material/CssBaseline";
 import Grid from "@mui/material/Grid";
 import Stack from "@mui/material/Stack";
 import Box from "@mui/material/Box";
-import Toolbar from "@mui/material/Toolbar";
+
 import Typography from "@mui/material/Typography";
 import Container from "@mui/material/Container";
-import Link from "@mui/material/Link";
+
 import { createTheme, ThemeProvider } from "@mui/material/styles";
-import Header from "./Header";
+
+
+import Header from "../components/Header";
 function DashBoard() {
   const cards = [1, 2, 3, 4, 5, 6, 7, 8, 9];
