+++ frontend/src/pages/EnvolopeFeedBack.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import "../../styles.css";
-import Header from '../Header';
+import "../styles.css";
+import Header from '../components/Header';
 
 export const EnvolopeFeedBack = () => {
   return (
