+++ frontend/src/pages/Meditation.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Header from '../Header'
+import Header from '../components/Header'
 export const Meditation = () => {
   return (
